move prop.toHash() (/core) to propertyHash() (/functions)
