--- packages/core/src/utils/hash-utils.ts
@@ -1,100 +1,0 @@
-import type { Ref, RefList, RefMap, RefSet } from 'property-graph';
-import type { Property } from '../properties/property.js';
-import { BufferUtils } from './buffer-utils.js';
-
-// Utilities for fast, non-cryptographic hashes on various types. Built for {@link Property.toHash}.
-// Deeply-equal inputs will always produce the same hash, but differing inputs may result in rare
-// hash collisions. Use {@link Property.equals} to confirm equality.
-//
-// TODO(perf): Hashes should be computed without heap allocations.
-// TODO(test): Verify that combining hashes with XOR is safe.
-
-export function hashRef(value: Ref<Property>, skip: Set<string>, cache: Map<Property, number>, depth: number): number {
-	const hashOptions = { skip, cache, depth: depth - 1 };
-	return depth > 0 ? value.getChild().toHash(hashOptions) : value.getChild().__id;
-}
-
-export function hashRefSet(
-	value: RefSet<Property> | RefList<Property>,
-	skip: Set<string>,
-	cache: Map<Property, number>,
-	depth: number,
-): number {
-	const hashOptions = { skip, cache, depth: depth - 1 };
-	let hash = 0;
-	for (const ref of value.values()) {
-		hash ^= depth > 0 ? ref.getChild().toHash(hashOptions) : ref.getChild().__id;
-	}
-	return hash;
-}
-
-export function hashRefMap(
-	value: RefMap<Property>,
-	skip: Set<string>,
-	cache: Map<Property, number>,
-	depth: number,
-): number {
-	const hashOptions = { skip, cache, depth: depth - 1 };
-	let hash = 0;
-	for (const key of value.keys()) {
-		const ref = value.get(key);
-		if (ref) {
-			hash ^= hashString(key);
-			hash ^= depth > 0 ? ref.getChild().toHash(hashOptions) : ref.getChild().__id;
-		}
-	}
-	return hash;
-}
-
-export function hashString(value: string): number {
-	return _hashBytes(BufferUtils.encodeText(value));
-}
-
-export function hashNumber(value: number): number {
-	return hashView(new Float64Array([value]));
-}
-
-export function hashBoolean(value: boolean): number {
-	return hashNumber(Number(value));
-}
-
-export function hashObject(value: object): number {
-	return hashString(JSON.stringify(value));
-}
-
-export function hashArray(value: unknown[]): number {
-	return hashString(JSON.stringify(value));
-}
-
-export function hashView(value: ArrayBufferView): number {
-	return _hashBytes(new Uint8Array(value.buffer, value.byteOffset, value.byteLength));
-}
-
-function _hashBytes(value: Uint8Array): number {
-	const bytes = BufferUtils.pad(value);
-	const bytesU32 = new Uint32Array(bytes.buffer, bytes.byteOffset, bytes.byteLength / Uint32Array.BYTES_PER_ELEMENT);
-	return murmurHash2(0, bytesU32);
-}
-
-/**
- * References:
- * - https://github.com/mikolalysenko/murmurhash-js/blob/f19136e9f9c17f8cddc216ca3d44ec7c5c502f60/murmurhash2_gc.js#L14
- * - https://github.com/zeux/meshoptimizer/blob/e47e1be6d3d9513153188216455bdbed40a206ef/src/indexgenerator.cpp#L12
- */
-function murmurHash2(h: number, key: Uint32Array): number {
-	const m = 0x5bd1e995;
-	const r = 24;
-
-	for (let i = 0, il = key.length; i < il; i++) {
-		let k = key[i];
-
-		k = Math.imul(k, m) >>> 0;
-		k = (k ^ (k >> r)) >>> 0;
-		k = Math.imul(k, m) >>> 0;
-
-		h = Math.imul(h, m) >>> 0;
-		h = (h ^ k) >>> 0;
-	}
-
-	return h;
-}
