+++ packages/core/src/utils/index.ts
@@ -4,8 +4,8 @@ export * from './file-utils.js';
 export * from './get-bounds.js';
 export * from './http-utils.js';
 export * from './image-utils.js';
-export * from './is-plain-object.js';
 export * from './logger.js';
 export * from './math-utils.js';
+export * from './object-utils.js';
 export * from './property-utils.js';
 export * from './uuid.js';
