+++ packages/core/test/properties/property.test.ts
@@ -83,67 +83,3 @@ test('listChildren', async (t) => {
 	t.deepEqual(graph.listChildren(nodeA), []);
 	t.deepEqual(graph.listChildren(nodeB), []);
 });
-
-test('toHash', async (t) => {
-	const document = new Document();
-	const materialA = document.createMaterial();
-	const materialB = document.createMaterial();
-
-	t.is(materialA.toHash(), materialB.toHash(), 'default == default');
-
-	materialA.setName('MaterialA');
-	materialB.setName('MaterialB');
-
-	t.not(materialA.toHash(), materialB.toHash(), 'name !== name');
-
-	const skip = new Set(['name']);
-	t.is(materialA.toHash({ skip }), materialB.toHash({ skip }), 'skip attributes');
-
-	materialA
-		.setBaseColorFactor([0.5, 0.5, 0.5, 1])
-		.setRoughnessFactor(1.0)
-		.setMetallicFactor(0)
-		.setAlphaMode('MASK')
-		.setDoubleSided(true)
-		.setExtras({ hello: 'world' });
-	materialB.copy(materialA);
-
-	t.is(materialA.toHash(), materialB.toHash(), 'primitive attributes, equal');
-
-	materialB.setDoubleSided(false);
-
-	t.not(materialA.toHash(), materialB.toHash(), 'primitive attributes, not equal');
-
-	const image = new Uint8Array([1, 2, 3, 4]);
-	const textureA = document.createTexture().setImage(image);
-	const textureB = document.createTexture().setImage(image);
-	materialB.copy(materialA);
-	materialA.setBaseColorTexture(textureA);
-	materialB.setBaseColorTexture(textureB);
-
-	t.is(materialA.toHash(), materialB.toHash(), 'ref attributes, equal');
-
-	textureB.setImage(new Uint8Array([5, 6, 7, 8]));
-
-	t.not(materialA.toHash(), materialB.toHash(), 'ref attributes, not equal');
-});
-
-test('toHash - depth', async (t) => {
-	// Use cases:
-	// - Skins may have deeply-equal joint nodes, but nodes are not interchangable.
-	// - Materials reference unique TextureInfos, which must be compared deeply.
-
-	const document = new Document();
-	const nodeA = document.createNode();
-	const nodeB = document.createNode().addChild(nodeA);
-	const nodeC = document.createNode().addChild(nodeB);
-
-	const nodeD = document.createNode();
-	const nodeE = document.createNode().addChild(nodeD);
-	const nodeF = document.createNode().addChild(nodeE);
-
-	t.true(nodeC.toHash({ depth: 0 }) !== nodeF.toHash({ depth: 0 }), 'depth = 0');
-	t.true(nodeC.toHash({ depth: 1 }) !== nodeF.toHash({ depth: 1 }), 'depth = 1');
-	t.true(nodeC.toHash({ depth: 2 }) === nodeF.toHash({ depth: 2 }), 'depth = 2');
-	t.true(nodeC.toHash({ depth: 3 }) === nodeF.toHash({ depth: 3 }), 'depth = 3');
-});
