+++ packages/functions/package.json
@@ -40,7 +40,8 @@
     "ktx-parse": "^1.0.1",
     "ndarray": "^1.0.19",
     "ndarray-lanczos": "^0.3.0",
-    "ndarray-pixels": "^5.0.1"
+    "ndarray-pixels": "^5.0.1",
+    "property-graph": "^4.0.0"
   },
   "files": [
     "dist/",
